app/requirements.txt
black==22.1.0
Django==4.0
djangorestframework==3.13.1
django-redis
redis
psycopg2-binary==2.9.2
celery
gunicorn
pytest-django==4.5.2
pytest==6.2.5
uvicorn[standard]
pytest-cov==3.0.0
flake8==4.0.1
isort==5.10.1
